+++ Changelog.md
@@ -13,6 +13,7 @@
 - Added CICD integration health test to check certificate results
 - Modified Kibana dashboards to show second region logs and metrics
 - Allow release branches to update image names using the kustomize image patch
+- Changed Slack channel for Argo notifications depending on IS_GA value
 
 _Changes:_
 
@@ -27,6 +28,7 @@ _Changes:_
 - [X] PDO-4569 Remove ES, Kibana and Grafana from second region
 - [X] PDO-4610 Retain set value for slack channel alerts
 - [X] PDO-4614 Automate pinning the branch for ping-cloud-dashboards in PCB
+- [X] PDO-4618 Default slack notifications using `IS_GA` env var
 - [X] PDO-4632 ALERT from the secondary region is shown as an ALERT from the primary region in the email message
 - [X] PDO-4648: Allow release branches to update image names using the kustomize image patch